[✨] 상품 찜, 찜한 목록, 키워드 추가  (#23)
* login

* 회원가입 진행중

* 휴면계정 진행중

* [:sparkles:] 마이페이지 상품

* [:recycle:] url 수정

* [:sparkles:] 내 상품 페이징

* [:truck:] resources 수정

* [:truck:] css 디렉토리 추가

* [:sparkles:] 내 상품 가격 표시

* [:art:] 프로필 사진 변경

* [:sparkles:] 상품 전체 개수 표시

* Delete log/2023-11/kimRestApiLog.2023-11-14.log

* 리스트 구현 중

* [:sparkles:] product list, detail

* [:bug:] git 충돌 수정

* [:recycle:] 사이드바 분리

* [:recycle:] 마이페이지 판매 상품

* [:art:] 마이페이지 키워드 추가

* [:recycle:] script fragment 분리

* [:recycle:] product insert의 fileupload 수정

* [:recycle:] 클래스 이름 수정

* [:sparkles:] 상품 찜

* [:bug:] sql 수정

* [:memo:] sql문 추가

* [:bug:] 검색 옵션 수정

* [:sparkles:] 마이페이지 찜 목록

* [:truck:] myshop 이동

* [:sparkles:] 키워드 등록

* [:recycle:] 코드 개선

---------

+++ src/main/java/com/team45/entity/FileData.java
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class FileData {
-    private int fileNo;
+    private Long fileNo;
     private String tableName;
-    private int columnNo;
+    private Long columnNo;
     private String originName;
     private String saveName;
     private String savePath;
